+++ app/Http/Controllers/FacilityController.php
@@ -18,78 +18,53 @@ class FacilityController extends Controller
         $this->facilityService = new FacilityService();
         $this->departmentService = new DepartmentService();
     }
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+  
     public function index()
     {
+        $this->authorize('viewAny',Facility::class);
         $facilities = $this->facilityService->facilityList();
         return view('facility.index',compact('facilities'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     * 
-     */
+ 
     public function create()
     {
+        $this->authorize('create',Facility::class);
         $departments = $this->departmentService->getDropdownList();
         return view('facility.create',compact('departments'));
     }
    
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+    
     public function store(FacilityRequest $request)
     {
+        $this->authorize('create',Facility::class);
         $validatedData = $request->validated(); 
         $this->facilityService->createOrUpdate($validatedData);
         Session::flash('message','Information stored successfully!!!!');
         return redirect()->route('facilities.index'); 
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Facility  $facility
-     * @return \Illuminate\Http\Response
-     */
+   
     public function show()
     {
         
         
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Facility  $facility
-     * @return \Illuminate\Http\Response
-     */
     public function edit($id)
     {
         $facility = $this->facilityService->getById($id);
+        $this->authorize('update',$facility);
         $departments = $this->departmentService->getDropdownList();            
         return view('facility.edit', compact('facility','departments'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Facility  $facility
-     * @return \Illuminate\Http\Response
-     */
+   
     public function update(FacilityRequest $request, $id)
     {
+        $facility = $this->facilityService->getById($id);
+        $this->authorize('update',$facility);
         
         $validatedData = $request->validated(); 
         $validatedData['id']=$id;
@@ -99,14 +74,11 @@ class FacilityController extends Controller
         return redirect()->route('facilities.index');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Facility  $facility
-     * @return \Illuminate\Http\Response
-     */
+   
     public function destroy($id)
     {
+        $facility = $this->facilityService->getById($id);
+        $this->authorize('delete',$facility);
         $facility = $this->facilityService->delete($id);
        
         if ($facility) {
