+++ resources/views/patient/index.blade.php
@@ -8,19 +8,32 @@
 <div class="br-pagebody">
    <div class="br-section-wrapper">
        @include('include._message')
-     <div class="row">
+       <div class="row">
          <div class="col-lg-6">
-           <h6 class="br-section-label">patients List</h6>
          </div>
+       
+            @can('create', App\Models\Patient::class)
+            <div class="col-lg-3 offset-lg-3">
+                <a href="{{route('patients.create')}}" class="btn btn-primary" style="margin-left:75px">Add New Patient</a>
+            </div>
+            @endcan
+       
+     
+       </div>
+       <br>
 
-
-
-         <div class="col-lg-3 offset-lg-3">
-             <a href="{{route('patients.create')}}" class="btn btn-primary" style="margin-left: 35px">Add New Patient</a>
-         </div>
+      
+      {!! Form::open(['url' => '/patient/live-search','method'=>'post']) !!}
+     <div class="row">
+        <div class="col-lg-2">
+            <button class="btn btn-primary">Search Patient</button>
+        </div>
+        <div class="col-lg-4">
+            {!! Form::text("query",null, ["class"=>"form-control form-control-label" ,"placeholder"=>"search patient by name, email or mobile"]) !!} 
+        </div>
      </div>
-
-   </br>
+     {!! Form::close() !!}
+     </br>
      
    <div class="table-responsive">
     <table class="table table-bordered table-colored table-dark">
@@ -35,6 +48,7 @@
                 <th>Age</th>
                 <th>Blood Group</th>
                 <th>Actions</th>
+                
             </tr>
 
 
@@ -50,17 +64,27 @@
               <td class="center">{{$patient->birthdate}}</td>
               <td class="center">{{$patient->age}}</td>
               <td class="center">{{$patient->blood->blood_group}}</td>
-                       
-              <td class="center">
                  
+              <td class="center">
+                @can('view',$patient)
+                  <a class="btn btn-info" href="{{route('patients.show',$patient->id)}}">
+                    <i class="fas fa-eye">View</i> 
+                  </a>
+                  @endcan
+
+                  @can('update',$patient)
                   <a class="btn btn-primary" href="{{route('patients.edit',$patient->id)}}">
                     <i class="fas fa-edit">Edit</i> 
                   </a>
+                  @endcan
+                  @can('delete',$patient)
                   
                   {!! Form::open(['method' => 'DELETE','route' => ['patients.destroy', $patient->id],'style'=>'display:inline']) !!}
                   {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
                   {!! Form::close() !!}
+                  @endcan
               </td>
+            
           </tr>
           @endforeach
         </tbody>
